Fix a formatting mismatch. (dm)
git-svn-id: svn://mielke.cc/main/brltty@6291 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/log.c
@@ -87,7 +87,7 @@ writeLogRecord (const char *record) {
       length = formatSeconds(buffer, sizeof(buffer), "%Y-%m-%d@%H:%M:%S", now.seconds);
       milliseconds = now.nanoseconds / NSECS_PER_MSEC;
 
-      fprintf(logFile, "%.*s.%03u ", length, buffer, milliseconds);
+      fprintf(logFile, "%.*s.%03u ", (int)length, buffer, milliseconds);
     }
 
     fputs(record, logFile);
